abstracted mdx functions
