+++ src/types/PostFile.ts
@@ -1,0 +1,9 @@
+import { GrayMatterFile } from 'gray-matter'
+
+interface PostFile {
+  filename: string
+  content: string
+  matter: GrayMatterFile<string>
+}
+
+export default PostFile
